# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartHome IoT API",
        "description": "API documentation for SmartHome IoT system",
        "contact": {
            "email": "leducanh1503.works@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/sensor-data": {
            "get": {
                "tags": [
                    "Sensor Data"
                ],
                "summary": "Get sensor data history",
                "description": "Retrieve the list of sensor data with pagination",
                "operationId": "getSensorDataHistory",
                "parameters": [
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "description": "List of sensor data",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "temperature": {
                                                        "type": "number",
                                                        "example": 25.5
                                                    },
                                                    "humidity": {
                                                        "type": "number",
                                                        "example": 60.3
                                                    },
                                                    "light": {
                                                        "type": "integer",
                                                        "example": 800
                                                    },
                                                    "received_at": {
                                                        "type": "string",
                                                        "example": "2023-10-05 12:34:56"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "description": "Pagination details",
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sensor-data/latest": {
            "get": {
                "tags": [
                    "Sensor Data"
                ],
                "summary": "Get the latest sensor data",
                "description": "Retrieve the latest sensor data with pagination",
                "operationId": "getLatestSensorData",
                "parameters": [
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "description": "Latest sensor data",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "temperature": {
                                                        "type": "number",
                                                        "example": 26
                                                    },
                                                    "humidity": {
                                                        "type": "number",
                                                        "example": 65
                                                    },
                                                    "light": {
                                                        "type": "integer",
                                                        "example": 700
                                                    },
                                                    "received_at": {
                                                        "type": "string",
                                                        "example": "2023-10-05 13:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sensor-data/filter": {
            "get": {
                "tags": [
                    "Sensor Data"
                ],
                "summary": "Filter sensor data",
                "description": "Retrieve filtered sensor data based on search criteria and sorting options",
                "operationId": "filterSensorData",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string for filtering by date, temperature, humidity, or light",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "25.5"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sorting order: 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "asc"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by: 'received_at', 'temperature', 'humidity', or 'light'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "received_at"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Field to filter by: 'temperature', 'humidity', 'light', or 'received_at'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "temperature",
                                "humidity",
                                "light",
                                "received_at"
                            ],
                            "example": "temperature"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Filtered sensor data",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "temperature": {
                                                        "type": "number",
                                                        "example": 25.5
                                                    },
                                                    "humidity": {
                                                        "type": "number",
                                                        "example": 60.3
                                                    },
                                                    "light": {
                                                        "type": "integer",
                                                        "example": 800
                                                    },
                                                    "received_at": {
                                                        "type": "string",
                                                        "example": "2023-10-05 12:34:56"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "description": "Pagination details",
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}